initing the basic of the break down for the webserver project
