+++ include/imports.hpp
@@ -1,4 +1,13 @@
 #pragma once
 
 #include <iostream>
-
+#include <string>
+#include <map>
+#include <vector>
+#include <sstream>
+#include <unistd.h>
+#include <fcntl.h>
+#include <sys/types.h>
+#include <sys/stat.h>
+#include <sys/socket.h>
+#include <netinet/in.h>
