+++ src/http/HttpRequest/HttpRequest.cpp
@@ -1,0 +1,20 @@
+#include "HttpRequest.hpp"
+
+HttpRequest::HttpRequest() : method(""), uri(""), version("") 
+{
+    std::cout << "HttpRequest constructor called" << std::endl;
+}
+HttpRequest::~HttpRequest() 
+{
+    std::cout << "HttpRequest destructor called" << std::endl;
+}
+void HttpRequest::parseRequest(const std::string &request) 
+{
+    // This is just to show that the function is called
+    (void)request;
+
+    method = "GET";
+    uri = "/index.html";
+    version = "HTTP/1.1";
+    std::cout << "Method: " << method << "\nURI: " << uri << "\nVersion: " << version << std::endl;
+}
