+++ src/http/HttpResponse/HttpResonse.cpp
@@ -1,0 +1,18 @@
+#include "./HttpResponse.hpp"
+
+HttpResponse::HttpResponse() : status(""), body("") 
+{
+    std::cout << "HttpResponse constructor called" << std::endl;
+}
+HttpResponse::~HttpResponse() 
+{
+    std::cout << "HttpResponse destructor called" << std::endl;
+}
+void HttpResponse::generateResponse(const std::string &status, const std::string &body)
+{
+    // This is just to show that the function is called
+    (void)status;
+    (void)body;
+    this->status = status;
+    std::cout << "Status: " << this->status << "\nBody: " << body << std::endl;
+}
