+++ src/main.cpp
@@ -1,14 +1,10 @@
 #include "../include/imports.hpp"
 #include "../include/enums.hpp"
-#include "./HttpRequest/HttpRequest.hpp"
+#include "http/HttpRequest/HttpRequest.hpp"
+#include "http/HttpResponse/HttpResponse.hpp"
 
 int main(void)
 {
-    // HttpRequest req;
-
-    // std::string request = "GET /index.html HTTP/1.1\r\nHost: example.com\r\n\r\n";
-    // std::cout << NOT_FOUND << std::endl;
-    // req.parseRequest(request);
-    // std::cout << "testing " << std::endl;
+    std::cout << "Web Server Starting..." << std::endl;
     return 0;
 }
