+++ src/network/Notes.md
@@ -1,0 +1,2 @@
+yuotsubo
+This is where all the networking classes and functions will go
